fix: remove shelljs and move context impl
