+++ packages/lightning-lsp-common/package.json
@@ -38,7 +38,6 @@
     "@types/ejs": "^3.1.5",
     "@types/jest": "^29.5.14",
     "@types/mock-fs": "^4.13.4",
-    "@types/shelljs": "^0.8.15",
     "@types/tmp": "^0.1.0",
     "find-node-modules": "^1.0.4",
     "jest": "^29.7.0",
@@ -46,7 +45,6 @@
     "lwc": "2.37.3",
     "mock-fs": "^5.5.0",
     "prettier": "^2.0.5",
-    "shelljs": "^0.8.5",
     "tmp": "^0.0.33",
     "ts-jest": "^29.2.6",
     "typescript": "5.0.4"
