+++ packages/lwc-language-server/package.json
@@ -40,7 +40,6 @@
     "comment-parser": "^0.7.6",
     "fast-glob": "^3.3.3",
     "normalize-path": "^3.0.0",
-    "shelljs": "^0.10.0",
     "vscode-html-languageservice": "^5.5.1",
     "vscode-languageserver": "^5.2.1",
     "vscode-uri": "^2.1.2",
